feat : create PastElectionsView basic layout and data
